refactor: updates historic prices layout
